autentifikacija i autorizacija uradjene, gotovo za 6 od 14 tabela

+++ ERP_backend/Controllers/VelicinaController.cs
@@ -2,10 +2,13 @@
 using ERP_backend.Entity;
 using ERP_backend.Model;
 using ERP_backend.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace ERP_backend.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("api/velicine")]
     [Produces("application/json", "application/xml")]
@@ -22,6 +25,7 @@ namespace ERP_backend.Controllers
             this.mapper = mapper;
         }
 
+        [AllowAnonymous]
         [HttpHead]
         [HttpGet]
         public ActionResult<List<VelicinaDTO>> GetAllVelicine()
@@ -43,13 +47,14 @@ namespace ERP_backend.Controllers
             }
         }
 
+        [AllowAnonymous]
         [HttpGet("{velicinaID}")]
         public ActionResult<VelicinaDTO> GetVelicinaByID(int velicinaID)
         {
             try
             {
                 VelicinaEntity? velicina = velicinaRepository.GetVelicinaByID(velicinaID);
-                if(velicina == null)
+                if (velicina == null)
                     return NoContent();
 
                 VelicinaDTO velicinaDTO = mapper.Map<VelicinaDTO>(velicina);
@@ -67,10 +72,24 @@ namespace ERP_backend.Controllers
         {
             try
             {
-                VelicinaDTO velicinaDTO = velicinaRepository.CreateVelicina(velicinaCreateDTO);
-                velicinaRepository.SaveChanges();
+                if (HttpContext.User.Claims.FirstOrDefault(e => e.Type == ClaimTypes.Role)?.Value == "ADMIN")
+                {
+                    List<VelicinaEntity> velicine = velicinaRepository.GetAllVelicine();
+                    if (velicine.Find(e => e.Oznaka == velicinaCreateDTO.Oznaka) == null)
+                    {
+                        VelicinaDTO velicinaDTO = velicinaRepository.CreateVelicina(velicinaCreateDTO);
+                        velicinaRepository.SaveChanges();
 
-                return Ok("Uspesno kreirana velicina!");
+                        return Ok("Uspesno kreirana velicina!");
+                    }
+                    else
+                    {
+                        return StatusCode(StatusCodes.Status422UnprocessableEntity, "Već postoji zadata oznaka velicine.");
+                    }
+                }
+                else
+                    return StatusCode(StatusCodes.Status403Forbidden, "Access forbiden");
+
             }
             catch (Exception ex)
             {
@@ -79,18 +98,52 @@ namespace ERP_backend.Controllers
         }
 
         [HttpDelete("{velicinaID}")]
-        public IActionResult DeleteVelicina (int velicinaID)
+        public IActionResult DeleteVelicina(int velicinaID)
         {
             try
             {
-                VelicinaEntity? velicina = velicinaRepository.GetVelicinaByID(velicinaID);
-                if (velicina == null)
+                if (HttpContext.User.Claims.FirstOrDefault(e => e.Type == ClaimTypes.Role)?.Value == "ADMIN")
+                {
+                    VelicinaEntity? velicina = velicinaRepository.GetVelicinaByID(velicinaID);
+                    if (velicina == null)
+                        return NotFound();
+
+                    velicinaRepository.DeleteVelicina(velicinaID);
+                    velicinaRepository.SaveChanges();
+
                     return NoContent();
+                }
+                else
+                    return StatusCode(StatusCodes.Status403Forbidden, "Access forbiden");
 
-                velicinaRepository.DeleteVelicina(velicinaID);
-                velicinaRepository.SaveChanges();
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
+        }
 
-                return NoContent();
+        [HttpPut]
+        [Consumes("application/json")]
+        public ActionResult<VelicinaDTO> UpdateVelicina([FromBody] VelicinaUpdateDTO velicinaUpdateDTO)
+        {
+            try
+            {
+                if (HttpContext.User.Claims.FirstOrDefault(e => e.Type == ClaimTypes.Role)?.Value == "ADMIN")
+                {
+                    VelicinaEntity? oldVelicina = velicinaRepository.GetVelicinaByID(velicinaUpdateDTO.IDVelicina);
+
+                    if (oldVelicina == null)
+                        return NotFound();
+
+                    VelicinaEntity velicina = mapper.Map<VelicinaEntity>(velicinaUpdateDTO);
+                    mapper.Map(velicina, oldVelicina);
+                    velicinaRepository.SaveChanges();
+                    return Ok(mapper.Map<VelicinaDTO>(velicina));
+                }
+                else
+                    return StatusCode(StatusCodes.Status403Forbidden, "Access forbiden");
+
             }
             catch (Exception ex)
             {
